Update learn_words.sh
Improve randomization by making the loop pass over the entire file (in a
shuffled order) first, after which it creates another permutation of the
input file and loops over it again. In addition, the urgency is now set
to critical so that the user can control the time before the answer
appears. Finally, an icon has been added to the popup.